backend/prediction_service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.0
aiohttp==3.9.1
redis==5.0.1
python-dotenv==1.0.0 